moved out initialization and phase methods
